Sum client side validation
The validation is made to match the server side validation as closely as possible. Only exception is negative numbers are not allowed on the client side. (I don't regard this to be important enough to be handled on the server side, but there's still no point in allowing it on the client side.