Fix upload button
CSS hack to hide the input element and style the label as a button. This makes the html validation error relating to the `required` attribute work as expected, while still making it possible to style the upload button. Previously the upload button was hidden with display: none, which resulted in a browser error and no validation error.