The entity_repo shows the parents in control_panel
All the entity-variables need to be translated to english. The delete function in the control_panel table does not work, since it does not have a path. Should this function be there or somewhere else? Via email? Also need to send out email as verification to the parents who sign up.