Fixed MazeMap integration with course location. Via link from admin.
Next time: make sloppy JS in frontend better. Also add capabitilites for
the admins to be able to change the url for the location on a later date
from kontrollpanel. Perhaps also mention the format of the "Skriv sted
for kurset", for some consistency, but this is maybe not so important.