Trying to fix posting to external_delete for deleting parents from
email-key. Getting 500 response when posting.
Next time: Try to make a button in dev which POSTs to the url, such that
I can get help/more information about what is wrong (since curl is not
giving me a lot of information in the response.)
Also fix the test. After this I think I can open the PR
(HacktoberFest;))